+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/BaseSparkJobIntegrationTest.java
@@ -14,8 +14,8 @@ import scala.Tuple2;
 
 public class BaseSparkJobIntegrationTest {
 
-  public static final String RESOURCE_PATH = "src/test/resources/wordcount.txt";
   protected SparkSession spark;
+
   protected SparkDataSource sut;
 
   protected JavaRDD<String> testFile;
@@ -29,6 +29,7 @@ public class BaseSparkJobIntegrationTest {
         .domain(Domain.SCIENTIFIC)
         .description("Yer a wizard harry")
         .events(Map.of("event1", "Desc of event1", "event2", "Desc of event2"))
+        .outputPath("src/test/resources/WTA")
         .build();
 
     SparkConf conf = new SparkConf()