+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/listener/TaskLevelListenerTest.java
@@ -17,7 +17,7 @@ import org.apache.spark.scheduler.TaskLocation;
 import org.apache.spark.storage.RDDInfo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import scala.collection.immutable.Seq;
+import scala.collection.Seq;
 import scala.collection.mutable.ListBuffer;
 
 class TaskLevelListenerTest extends BaseLevelListenerTest {
@@ -55,17 +55,17 @@ class TaskLevelListenerTest extends BaseLevelListenerTest {
   @Test
   void testTaskEndMetricExtraction() {
     ListBuffer<StageInfo> stageBuffer = new ListBuffer<>();
-    stageBuffer.addOne(testStageInfo);
+    stageBuffer.$plus$eq(testStageInfo);
 
     fakeTaskListener.onJobStart(new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties()));
     fakeTaskListener.onTaskEnd(taskEndEvent);
     assertEquals(1, fakeTaskListener.getProcessedObjects().size());
     Task curTask = fakeTaskListener.getProcessedObjects().get(0);
-    assertEquals(1, curTask.getId());
+    assertEquals(2, curTask.getId());
     assertEquals("testtaskType", curTask.getType());
     assertEquals(50L, curTask.getSubmitTime());
     assertEquals(100L, curTask.getRuntime());
-    assertEquals(1L, curTask.getWorkflowId());
+    assertEquals(2L, curTask.getWorkflowId());
     assertEquals("testUser".hashCode(), curTask.getUserId());
     assertEquals(-1, curTask.getSubmissionSite());
     assertEquals("N/A", curTask.getResourceType());