+++ core/src/main/java/com/asml/apa/wta/core/config/RuntimeConfig.java
@@ -9,6 +9,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+/**
+ * Config class for the plugin.
+ *
+ * @author Henry Page
+ * @author Lohithsai Yadala Chanchu
+ * @author Pil Kyu Cho
+ * @author Atour Mousavi Gourabi
+ * @since 1.0.0
+ */
 @Getter
 @Setter
 @Builder
@@ -28,4 +37,15 @@ public class RuntimeConfig {
 
   @Builder.Default
   private String logLevel = "ERROR";
+
+  @Builder.Default
+  private boolean isStageLevel = false;
+
+  @Builder.Default
+  private int resourcePingInterval = 1000;
+
+  @Builder.Default
+  private int executorSynchronizationInterval = 2000;
+
+  private String outputPath;
 }