+++ core/src/main/java/com/asml/apa/wta/core/logger/Log4j2Configuration.java
@@ -3,7 +3,16 @@ package com.asml.apa.wta.core.logger;
 import com.asml.apa.wta.core.config.RuntimeConfig;
 import com.asml.apa.wta.core.utils.WtaUtils;
 import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.core.appender.rolling.action.Duration;
+import org.apache.logging.log4j.core.config.Configuration;
 import org.apache.logging.log4j.core.config.Configurator;
+import org.apache.logging.log4j.core.config.builder.api.AppenderComponentBuilder;
+import org.apache.logging.log4j.core.config.builder.api.ComponentBuilder;
+import org.apache.logging.log4j.core.config.builder.api.ConfigurationBuilder;
+import org.apache.logging.log4j.core.config.builder.api.ConfigurationBuilderFactory;
+import org.apache.logging.log4j.core.config.builder.api.LayoutComponentBuilder;
+import org.apache.logging.log4j.core.config.builder.api.RootLoggerComponentBuilder;
+import org.apache.logging.log4j.core.config.builder.impl.BuiltConfiguration;
 
 /**
  * Logging utils to set up part of the Log4j2 configuration.
@@ -22,6 +31,53 @@ public class Log4j2Configuration {
    */
   private Log4j2Configuration() {
     throw new IllegalStateException();
+  }
+
+  /**
+   * Sets up the logging configuration.
+   *
+   * @param logLevel the logging level to be used
+   * @author Atour Mousavi Gourabi
+   * @since 1.0.0
+   */
+  public static void setUpLoggingConfig(Level logLevel) {
+    ConfigurationBuilder<BuiltConfiguration> builder = ConfigurationBuilderFactory.newConfigurationBuilder();
+
+    LayoutComponentBuilder layoutBuilder = builder.newLayout("PatternLayout")
+        .addAttribute("pattern", "%d{yyyy-MM-dd HH:mm:ss.SSS} [%-5p] [%t] %C{3}.%M(%F:%L) - %m%n");
+    AppenderComponentBuilder consoleAppenderBuilder =
+        builder.newAppender("consoleLogger", "console").add(layoutBuilder);
+    ComponentBuilder<?> policyBuilder = builder.newComponent("Policies")
+        .addComponent(builder.newComponent("TimeBasedTriggeringPolicy")
+            .addAttribute("interval", 1)
+            .addAttribute("modulate", true))
+        .addComponent(builder.newComponent("SizeBasedTriggeringPolicy").addAttribute("size", "10MB"));
+    ComponentBuilder<?> strategyBuilder = builder.newComponent("DefaultRolloverStrategy")
+        .addAttribute("max", 10)
+        .addAttribute("min", 1)
+        .addAttribute("fileIndex", "min")
+        .addComponent(builder.newComponent("Delete")
+            .addAttribute("basePath", "log")
+            .addAttribute("maxDepth", 10)
+            .addComponent(
+                builder.newComponent("IfLastModified").addAttribute("age", Duration.parse("365d"))));
+    AppenderComponentBuilder fileAppenderBuilder = builder.newAppender("fileLogger", "RollingFile")
+        .addAttribute("fileName", "logs/wta.log")
+        .addAttribute("filePattern", "logs/wta_%d{yyyyMMdd}_%i.log")
+        .add(layoutBuilder)
+        .addComponent(policyBuilder)
+        .addComponent(strategyBuilder);
+    builder.add(consoleAppenderBuilder).add(fileAppenderBuilder);
+
+    RootLoggerComponentBuilder consoleComponent = builder.newRootLogger(logLevel)
+        .add(builder.newAppenderRef("consoleLogger"))
+        .add(builder.newAppenderRef("fileLogger"));
+    builder.add(consoleComponent);
+
+    Configuration configuration = builder.build();
+
+    Configurator.initialize(configuration);
+    Configurator.setRootLevel(logLevel);
   }
 
   /**
@@ -61,6 +117,6 @@ public class Log4j2Configuration {
         logLevel = Level.ERROR;
         break;
     }
-    Configurator.setRootLevel(logLevel);
+    setUpLoggingConfig(logLevel);
   }
 }