+++ core/src/main/java/com/asml/apa/wta/core/model/BaseTraceObject.java
@@ -1,6 +1,8 @@
 package com.asml.apa.wta.core.model;
 
 import java.io.Serializable;
+import org.apache.avro.Schema;
+import org.apache.avro.generic.GenericRecord;
 
 /**
  * BaseSchema interface that all schema objects implement.
@@ -9,6 +11,7 @@ import java.io.Serializable;
  * @since 1.0.0
  */
 public interface BaseTraceObject extends Serializable {
+
   /**
    * Returns a hardcoded schema version.
    *
@@ -19,4 +22,14 @@ public interface BaseTraceObject extends Serializable {
   default String getSchemaVersion() {
     return "1.0";
   }
+
+  /**
+   * All WTA objects that are stored as Parquet files rely on this method to convert the object to a record.
+   * It should build the record object based on the checker and the schema provided.
+   *
+   * @param checker array indicating what column to ignore
+   * @param schema schema for the output object
+   * @return record of the object
+   */
+  GenericRecord convertToRecord(Boolean[] checker, Schema schema);
 }