+++ core/src/main/java/com/asml/apa/wta/core/utils/ParquetWriterUtils.java
@@ -4,15 +4,18 @@ import com.asml.apa.wta.core.model.Resource;
 import com.asml.apa.wta.core.model.Task;
 import com.asml.apa.wta.core.model.Workflow;
 import com.asml.apa.wta.core.model.Workload;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.ByteArrayOutputStream;
+import com.google.gson.FieldNamingPolicy;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.OutputStream;
+import java.io.FileWriter;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.avro.Schema;
 import org.apache.avro.SchemaBuilder;
 import org.apache.avro.generic.GenericRecord;
@@ -25,6 +28,7 @@ import org.apache.avro.generic.GenericRecord;
  */
 @SuppressWarnings({"CyclomaticComplexity", "HiddenField"})
 @Getter
+@Slf4j
 public class ParquetWriterUtils {
 
   private final String version;
@@ -39,12 +43,19 @@ public class ParquetWriterUtils {
 
   private Workload workload = null;
 
+  private final Map<String, String> files;
+
   public ParquetWriterUtils(File path, String version) {
     resources = new ArrayList<>();
     tasks = new ArrayList<>();
     workflows = new ArrayList<>();
     this.path = path;
     this.version = version;
+    files = Map.of(
+        "resources", "resource.parquet",
+        "tasks", "task.parquet",
+        "workflows", "workflow.parquet",
+        "workload", "generic_information.json");
   }
 
   /**
@@ -98,8 +109,9 @@ public class ParquetWriterUtils {
    * @param taskFileName task file name
    * @param workflowFileName workflow file name
    * @param workloadFileName workload file name
-   * @throws Exception possible exception due to io
+   * @throws Exception possible exception due to I/O error
    * @since 1.0.0
+   * @author Pil Kyu Cho
    * @author Tianchen Qu
    */
   public void writeToFile(
@@ -132,20 +144,24 @@ public class ParquetWriterUtils {
           fieldSchema.name("type").type().nullable().stringType().stringDefault("test");
     }
     if (checker[2]) {
-      fieldSchema = fieldSchema.name("numResources").type().doubleType().doubleDefault(0.0);
+      fieldSchema = fieldSchema.name("num_resources").type().doubleType().doubleDefault(0.0);
     }
     if (checker[3]) {
-      fieldSchema =
-          fieldSchema.name("procModel").type().nullable().stringType().stringDefault("test");
+      fieldSchema = fieldSchema
+          .name("proc_model")
+          .type()
+          .nullable()
+          .stringType()
+          .stringDefault("test");
     }
     if (checker[4]) {
       fieldSchema = fieldSchema.name("memory").type().longType().longDefault(0);
     }
     if (checker[5]) {
-      fieldSchema = fieldSchema.name("diskSpace").type().longType().longDefault(0);
+      fieldSchema = fieldSchema.name("disk_space").type().longType().longDefault(0);
     }
     if (checker[6]) {
-      fieldSchema = fieldSchema.name("networkSpeed").type().longType().longDefault(0);
+      fieldSchema = fieldSchema.name("network").type().longType().longDefault(0);
     }
     if (checker[7]) {
       fieldSchema = fieldSchema.name("os").type().nullable().stringType().stringDefault("test");
@@ -155,11 +171,15 @@ public class ParquetWriterUtils {
           fieldSchema.name("details").type().nullable().stringType().stringDefault("test");
     }
     Schema schema = fieldSchema.endRecord();
-    AvroUtils resourceWriter =
-        new AvroUtils(schema, new File(path, "/resources/" + version + "/" + resourceFileName + ".parquet"));
+    AvroUtils resourceWriter = new AvroUtils(
+        schema,
+        new File(
+            path,
+            Paths.get("resources", version, resourceFileName + ".parquet")
+                .toString()));
     List<GenericRecord> resourceList = new ArrayList<>();
     for (Resource resource : resources) {
-      resourceList.add(Resource.convertResourceToRecord(resource, checker, schema));
+      resourceList.add(resource.convertToRecord(checker, schema));
     }
     resourceWriter.writeRecords(resourceList);
     resourceWriter.close();
@@ -175,7 +195,7 @@ public class ParquetWriterUtils {
    */
   private void writeTaskToFile(String taskFileName) throws Exception {
     Boolean[] checker = checkTaskDomain(tasks);
-    SchemaBuilder.FieldAssembler<Schema> fieldSchema = SchemaBuilder.record("resource")
+    SchemaBuilder.FieldAssembler<Schema> fieldSchema = SchemaBuilder.record("task")
         .namespace("com.asml.apa.wta.core.model")
         .fields();
     if (checker[0]) {
@@ -186,17 +206,17 @@ public class ParquetWriterUtils {
           fieldSchema.name("type").type().nullable().stringType().stringDefault("test");
     }
     if (checker[2]) {
-      fieldSchema = fieldSchema.name("submitTime").type().longType().longDefault(0);
+      fieldSchema = fieldSchema.name("ts_submit").type().longType().longDefault(0);
     }
     if (checker[3]) {
-      fieldSchema = fieldSchema.name("submissionSite").type().intType().intDefault(0);
+      fieldSchema = fieldSchema.name("submission_site").type().intType().intDefault(0);
     }
     if (checker[4]) {
       fieldSchema = fieldSchema.name("runtime").type().longType().longDefault(0);
     }
     if (checker[5]) {
       fieldSchema = fieldSchema
-          .name("resourceType")
+          .name("resource_type")
           .type()
           .nullable()
           .stringType()
@@ -204,7 +224,7 @@ public class ParquetWriterUtils {
     }
     if (checker[6]) {
       fieldSchema = fieldSchema
-          .name("resourceAmountRequested")
+          .name("resource_amount_requested")
           .type()
           .doubleType()
           .doubleDefault(0.0);
@@ -230,20 +250,20 @@ public class ParquetWriterUtils {
           .arrayDefault(new ArrayList<Long>());
     }
     if (checker[9]) {
-      fieldSchema = fieldSchema.name("userId").type().intType().intDefault(0);
+      fieldSchema = fieldSchema.name("user_id").type().intType().intDefault(0);
     }
     if (checker[10]) {
-      fieldSchema = fieldSchema.name("groupId").type().intType().intDefault(0);
+      fieldSchema = fieldSchema.name("group_id").type().intType().intDefault(0);
     }
     if (checker[11]) {
       fieldSchema =
           fieldSchema.name("nfrs").type().nullable().stringType().stringDefault("test");
     }
     if (checker[12]) {
-      fieldSchema = fieldSchema.name("workflowId").type().longType().longDefault(0);
+      fieldSchema = fieldSchema.name("workflow_id").type().longType().longDefault(0);
     }
     if (checker[13]) {
-      fieldSchema = fieldSchema.name("waitTime").type().longType().longDefault(0);
+      fieldSchema = fieldSchema.name("wait_time").type().longType().longDefault(0);
     }
     if (checker[14]) {
       fieldSchema =
@@ -251,31 +271,34 @@ public class ParquetWriterUtils {
     }
     if (checker[15]) {
       fieldSchema =
-          fieldSchema.name("memoryRequested").type().doubleType().doubleDefault(0.0);
+          fieldSchema.name("memory_requested").type().doubleType().doubleDefault(0.0);
     }
     if (checker[16]) {
-      fieldSchema = fieldSchema.name("networkIoTime").type().longType().longDefault(0);
+      fieldSchema = fieldSchema.name("network_io_time").type().longType().longDefault(0);
     }
     if (checker[17]) {
-      fieldSchema = fieldSchema.name("diskIoTime").type().longType().longDefault(0);
+      fieldSchema = fieldSchema.name("disk_io_time").type().longType().longDefault(0);
     }
     if (checker[18]) {
       fieldSchema =
-          fieldSchema.name("diskSpaceRequested").type().doubleType().doubleDefault(0.0);
+          fieldSchema.name("disk_space_requested").type().doubleType().doubleDefault(0.0);
     }
     if (checker[19]) {
       fieldSchema =
-          fieldSchema.name("energyConsumption").type().longType().longDefault(0);
+          fieldSchema.name("energy_consumption").type().doubleType().doubleDefault(0.0);
     }
     if (checker[20]) {
-      fieldSchema = fieldSchema.name("resourceUsed").type().longType().longDefault(0);
+      fieldSchema = fieldSchema.name("resource_used").type().longType().longDefault(0);
     }
     Schema schema = fieldSchema.endRecord();
-    AvroUtils taskWriter =
-        new AvroUtils(schema, new File(path, "/tasks/" + version + "/" + taskFileName + ".parquet"));
+    AvroUtils taskWriter = new AvroUtils(
+        schema,
+        new File(
+            path,
+            Paths.get("tasks", version, taskFileName + ".parquet").toString()));
     List<GenericRecord> taskList = new ArrayList<>();
     for (Task task : tasks) {
-      taskList.add(Task.convertTaskToRecord(task, checker, schema));
+      taskList.add(task.convertToRecord(checker, schema));
     }
     taskWriter.writeRecords(taskList);
     taskWriter.close();
@@ -291,14 +314,14 @@ public class ParquetWriterUtils {
    */
   private void writeWorkflowToFile(String workflowFileName) throws Exception {
     Boolean[] checker = checkWorkflowDomain(workflows);
-    SchemaBuilder.FieldAssembler<Schema> fieldSchema = SchemaBuilder.record("resource")
+    SchemaBuilder.FieldAssembler<Schema> fieldSchema = SchemaBuilder.record("workflow")
         .namespace("com.asml.apa.wta.core.model")
         .fields();
     if (checker[0]) {
       fieldSchema = fieldSchema.name("id").type().longType().noDefault();
     }
     if (checker[1]) {
-      fieldSchema = fieldSchema.name("submitTime").type().longType().noDefault();
+      fieldSchema = fieldSchema.name("ts_submit").type().longType().noDefault();
     }
     if (checker[2]) {
       fieldSchema = fieldSchema
@@ -311,22 +334,22 @@ public class ParquetWriterUtils {
           .noDefault();
     }
     if (checker[3]) {
-      fieldSchema = fieldSchema.name("numberOfTasks").type().intType().noDefault();
+      fieldSchema = fieldSchema.name("task_count").type().intType().noDefault();
     }
     if (checker[4]) {
       fieldSchema =
-          fieldSchema.name("criticalPathLength").type().intType().noDefault();
+          fieldSchema.name("critical_path_length").type().longType().noDefault();
     }
     if (checker[5]) {
-      fieldSchema =
-          fieldSchema.name("criticalPathTaskCount").type().intType().noDefault();
-    }
-    if (checker[6]) {
       fieldSchema = fieldSchema
-          .name("maxNumberOfConcurrentTasks")
+          .name("critical_path_task_count")
           .type()
           .intType()
           .noDefault();
+    }
+    if (checker[6]) {
+      fieldSchema =
+          fieldSchema.name("max_concurrent_tasks").type().intType().noDefault();
     }
     if (checker[7]) {
       fieldSchema =
@@ -342,7 +365,7 @@ public class ParquetWriterUtils {
     }
     if (checker[10]) {
       fieldSchema = fieldSchema
-          .name("applicationName")
+          .name("application_name")
           .type()
           .nullable()
           .stringType()
@@ -350,37 +373,48 @@ public class ParquetWriterUtils {
     }
     if (checker[11]) {
       fieldSchema = fieldSchema
-          .name("applicationField")
+          .name("application_field")
           .type()
           .nullable()
           .stringType()
           .noDefault();
     }
     if (checker[12]) {
-      fieldSchema = fieldSchema.name("totalResources").type().doubleType().noDefault();
+      fieldSchema =
+          fieldSchema.name("total_resources").type().doubleType().noDefault();
     }
     if (checker[13]) {
       fieldSchema =
-          fieldSchema.name("totalMemoryUsage").type().doubleType().noDefault();
+          fieldSchema.name("total_memory_usage").type().doubleType().noDefault();
     }
     if (checker[14]) {
       fieldSchema =
-          fieldSchema.name("totalNetworkUsage").type().longType().noDefault();
+          fieldSchema.name("total_network_usage").type().longType().noDefault();
     }
     if (checker[15]) {
-      fieldSchema =
-          fieldSchema.name("totalDiskSpaceUsage").type().longType().noDefault();
+      fieldSchema = fieldSchema
+          .name("total_disk_space_usage")
+          .type()
+          .doubleType()
+          .noDefault();
     }
     if (checker[16]) {
-      fieldSchema =
-          fieldSchema.name("totalEnergyConsumption").type().longType().noDefault();
+      fieldSchema = fieldSchema
+          .name("total_energy_consumption")
+          .type()
+          .doubleType()
+          .noDefault();
     }
     Schema schema = fieldSchema.endRecord();
-    AvroUtils workflowWriter =
-        new AvroUtils(schema, new File(path, "/workflows/" + version + "/" + workflowFileName + ".parquet"));
+    AvroUtils workflowWriter = new AvroUtils(
+        schema,
+        new File(
+            path,
+            Paths.get("workflows", version, workflowFileName + ".parquet")
+                .toString()));
     List<GenericRecord> workflowList = new ArrayList<>();
     for (Workflow workflow : workflows) {
-      workflowList.add(Workflow.convertWorkflowToRecord(workflow, checker, schema));
+      workflowList.add(workflow.convertToRecord(checker, schema));
     }
     workflowWriter.writeRecords(workflowList);
     workflowWriter.close();
@@ -395,15 +429,14 @@ public class ParquetWriterUtils {
    * @author Tianchen Qu
    */
   private void writeWorkloadToFile(String workloadFileName) throws Exception {
-    ObjectMapper mapper = new ObjectMapper();
-    String workloadJson = mapper.writeValueAsString(workload);
-    File file = new File(this.path, "/workload/" + version);
+    File file = new File(Paths.get(this.path.getPath(), "workload", version).toString());
     file.mkdirs();
-    ByteArrayOutputStream writer = new ByteArrayOutputStream();
-    OutputStream dir = new FileOutputStream(new File(file, workloadFileName + ".json"));
-    writer.write(workloadJson.getBytes());
-    writer.writeTo(dir);
-    writer.flush();
+    Gson gson = new GsonBuilder()
+        .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
+        .create();
+    try (FileWriter fw = new FileWriter(new File(file, workloadFileName + ".json"))) {
+      gson.toJson(workload, fw);
+    }
   }
 
   /**
@@ -514,7 +547,7 @@ public class ParquetWriterUtils {
       if (task.getDiskSpaceRequested() == -1.0) {
         flg[18] = false;
       }
-      if (task.getEnergyConsumption() == -1) {
+      if (task.getEnergyConsumption() == -1.0) {
         flg[19] = false;
       }
       if (task.getResourceUsed() == -1) {
@@ -546,7 +579,7 @@ public class ParquetWriterUtils {
       if (workflow.getNumberOfTasks() == -1) {
         flg[3] = false;
       }
-      if (workflow.getCriticalPathLength() == -1) {
+      if (workflow.getCriticalPathLength() == -1L) {
         flg[4] = false;
       }
       if (workflow.getCriticalPathTaskCount() == -1) {
@@ -561,7 +594,7 @@ public class ParquetWriterUtils {
       if (workflow.getScheduler() == null) {
         flg[8] = false;
       }
-      if (workflow.getDomain() == null || workflow.getDomain().getValue() == null) {
+      if (workflow.getDomain() == null) {
         flg[9] = false;
       }
       if (workflow.getApplicationName() == null) {
@@ -582,10 +615,22 @@ public class ParquetWriterUtils {
       if (workflow.getTotalDiskSpaceUsage() == -1) {
         flg[15] = false;
       }
-      if (workflow.getTotalEnergyConsumption() == -1) {
+      if (workflow.getTotalEnergyConsumption() == -1.0) {
         flg[16] = false;
       }
     }
     return flg;
+  }
+
+  /**
+   * Deletes any potentially pre-existing parquet files. If files already exist, parquet writer
+   * will throw an exception.
+   * @author Pil Kyu Cho
+   * @since 1.0.0
+   */
+  public void deletePreExistingFiles() {
+    files.entrySet().stream().forEach(e -> new File(Paths.get(path.getPath(), e.getKey(), version, e.getValue())
+            .toString())
+        .delete());
   }
 }