+++ core/src/test/java/com/asml/apa/wta/core/RuntimeConfigTest.java
@@ -13,8 +13,12 @@ class RuntimeConfigTest {
     RuntimeConfig cr = RuntimeConfig.builder().build();
     assertThat(cr.getAuthors()).isNull();
     assertThat(cr.getDomain()).isNull();
+    assertThat(cr.isStageLevel()).isFalse();
     assertThat(cr.getDescription()).isEqualTo("");
     assertThat(cr.getEvents()).isEqualTo(new HashMap<String, String>());
     assertThat(cr.getLogLevel()).isEqualTo("ERROR");
+    assertThat(cr.getResourcePingInterval()).isEqualTo(1000);
+    assertThat(cr.getExecutorSynchronizationInterval()).isEqualTo(2000);
+    assertThat(cr.getOutputPath()).isNull();
   }
 }